requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
Automat==20.2.0
black==21.5b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect==1.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.44.0
django-cors-headers==3.5.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.0
djoser==2.1.0
docopt==0.6.2
gunicorn==20.0.4
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
lxml==4.6.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.0
parsel==1.6.0
pathspec==0.8.1
Pillow==8.2.0
pipreqs==0.4.10
priority==1.3.0
Protego==0.1.16
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dotenv==0.15.0
python-scrapyd-api==2.1.2
python3-openid==3.2.0
pytz==2021.1
queuelib==1.6.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
Scrapy==2.5.0
scrapy-user-agents==0.1.1
scrapyd==1.2.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.4
user-agents==2.2.0
w3lib==1.22.0
whitenoise==5.2.0
yarg==0.1.9
zope.interface==5.4.0
